# Perf/lazy insert ptr removal