This is to open the conversation. I haven't looked at the strict version or any other alternatives like `insertR`.

I'm not convinced that the optimisations for identical pointers is worth the complexity, and it hurts performance for inserts of new elements and the replacement of elements compared to this version.


```
Old
  insert absent:    OK (1.46s)
    352  μs ±  27 μs
  insert same:      OK (1.63s)
    194  μs ±  16 μs
  insert alternate: OK (1.45s)
    351  μs ±  28 μs


New
  insert absent:    OK (1.27s)
    312  μs ±  29 μs
  insert same:      OK (2.00s)
    242  μs ±  13 μs
  insert alternate: OK (1.27s)
    308  μs ±  29 μs
```