@treeowl 2024-01-12
I'm surprised you're getting good results with explicit continuations. Do you know how GHC is transforming those?