@HuwCampbell 2024-01-12
Yeah I couldn't really see any differences in the union benchmarks (maybe one was marginally faster?). From trying an explicit continuation on the other inserts, I expect that it is slower when there is a value inserted.

I haven't looked at the core yet.

I also considered just throwing a sentinel exception if `EQ` is found, then catching and returning the top map. It would probably be fast, but exceptions for control flow are a bit gross.