@treeowl 2024-01-12
I would expect exceptions to be slow. I think I might have tried? I don't
remember for sure. If you try, use unsafeDupablePerformIO to catch the
exception. Also consider using raise# and catch# manually, to allow an
exception type other than SomeException (this may require
reallyUnsafePtrEquality# to determine whether you got the not-present
exception).

On Thu, Jan 11, 2024, 10:31 PM Huw Campbell ***@***.***>
wrote:

> Yeah I couldn't really see any differences in the union benchmarks (maybe
> one was marginally faster?). From trying an explicit continuation on the
> other inserts, I expect that it is slower when there is a value inserted.
>
> I haven't looked at the core yet.
>
> I also considered just throwing a sentinel exception if EQ is found, then
> catching and returning the top map. It would probably be fast, but
> exceptions for control flow are a bit gross.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/haskell/containers/pull/980#issuecomment-1888378630>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAOOF7I45DIO6Q6DACGP4PLYOCVARAVCNFSM6AAAAABBWC3ULWVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMYTQOBYGM3TQNRTGA>
> .
> You are receiving this because you commented.Message ID:
> ***@***.***>
>