@HuwCampbell 2024-01-13
I'm not as concerned with `insertR` as `insert` and `insertWith`.

I guess the question is do we want a 10-15% sacrifice on insertion when the key in not already present but a 20-30% improvement when it is?

How would that decision be made?