@meooow25 2024-10-06
> I'm not convinced that the (unsafe) optimisations for identical pointers is worth the complexity, and it hurts performance for inserts of new elements and the replacement of elements compared to this version.
 
> So inserting when the key is missing (common) or a new item at the same key (also common), are faster, while only literally inserting the same identical item (i.e., a no-op, probably less common) is slower.

This sounds reasonable. The current code seems optimized for a very rare case. We have a comment acknowledging the same: https://github.com/haskell/containers/blob/41783ed917e5a16f3025728bcdb3148e73ff723d/containers/src/Data/Map/Internal.hs#L785-L788

But why is it likely in `union`? It would happen when `union`ing two maps with a large overlap in shared entries. That again seems like a rare situation.

---

There are a few variants of `insert` we can try:

A. Pointer equality (current)
B. Return a bool which controls rebalancing (this PR)
C. No tricks
D. ~~Both A and B~~ (let's not go there)

Curiously, strict map uses C instead of A: https://github.com/haskell/containers/blob/41783ed917e5a16f3025728bcdb3148e73ff723d/containers/src/Data/Map/Strict/Internal.hs#L503-L506

And there are 3 situations that can occur:

1. It's a new key. C is the best here, A and B waste work.
2. It's an existing key but the value is new. B is the best here. How likely is this? Perhaps not too unlikely, shared key objects might be around.
3. It's an existing key and existing value. C is the best here, and B should also help. This seems rather unlikely.

---

So, there's a bunch of benchmarking involved, if you want to figure this out.  
For `union` benchmarks you can use the `set-operations-map`, it covers various sets of inputs.  
Related to the history of the current code: #315, #416. Though I don't see any new information that is not already documented.